Moved exceptions to modules of associated classes.
