+++ src/algorithms/clustering/dbscan.py
@@ -6,7 +6,7 @@ from numpy import ndarray
 from numba import cuda
 from src.algorithms.clustering.dbscan_cpu import _dbscan_cpu
 from src.algorithms.clustering.dbscan_gpu import _dbscan_gpu
-from src.utilities.exception import InvalidEncodingError
+from src.utilities.utils import InvalidEncodingError
 
 
 class ClusterCategory(Enum):
