+++ src/algorithms/field_transforms.py
@@ -5,9 +5,8 @@ from random import Random
 from numpy import ndarray, array, dot, zeros
 from src.algorithms.channelling import load_crit_data, fraction
 from src.algorithms.clustering.dbscan import dbscan
-from src.data_structures.field import FieldLike, FieldType, Field
-from src.utilities.exception import FieldNullError, SymmetryNotImplementedError
-from src.data_structures.phase import Phase, CrystalFamily
+from src.data_structures.field import FieldLike, FieldType, Field, FieldNullError
+from src.data_structures.phase import Phase, CrystalFamily, SymmetryNotImplementedError
 from src.utilities.geometry import (
     Axis,
     AxisSet,
